HORTI - CREATE DOMAIN AND QUERY VALIDATION

+++ ServiceCoreHortiCommand/SERVICE/ProductDomainService.cs
@@ -1,7 +1,9 @@
 ﻿using CROSSCUTTINGCOREHORTI.EXTENSION;
 using DATAACCESSCOREHORTICOMMAND.ICOMMAND;
 using DOMAINCOREHORTICOMMAND;
+using FluentValidation;
 using SERVICECOREHORTICOMMAND.ISERVICE;
+using System;
 using System.Threading.Tasks;
 using VALIDATIONCOREHORTICOMMAND.DOMAIN;
 
@@ -9,31 +11,39 @@ namespace SERVICECOREHORTICOMMAND.SERVICE
 {
     public class ProductDomainService : IProductDomainService
     {
-        private readonly ProductDomainServiceValidation _productDomainServiceValidation;
+        private readonly CreateProductDomainServiceValidation _createProductDomainServiceValidation;
+        private readonly UpdateProductDomainServiceValidation _updateProductDomainServiceValidation;
+
         private readonly IProductRepository _productRepository;
 
-        public ProductDomainService(ProductDomainServiceValidation productDomainServiceValidation,
+        public ProductDomainService(CreateProductDomainServiceValidation createProductDomainServiceValidation,
+                                    UpdateProductDomainServiceValidation updateProductDomainServiceValidation,
                                     IProductRepository productRepository)
         {
+            _createProductDomainServiceValidation = createProductDomainServiceValidation;
+            _updateProductDomainServiceValidation = updateProductDomainServiceValidation;
             _productRepository = productRepository;
-            _productDomainServiceValidation = productDomainServiceValidation;
+            
         }
 
         public async Task ProductServiceCreate(Product product)
         {
-            _productDomainServiceValidation.ValidateHorti(product);
+            _createProductDomainServiceValidation.ValidateHorti(product);
 
             await _productRepository.CreateProduct(product);
         }
 
         public async Task ProductServiceDelete(Product product)
-        { 
+        {
+            if (product.IdProduct == Guid.Empty)
+                throw new ValidationException("PRODUCT NOT EXISTS!");
+
             await _productRepository.DeleteProduct(product);
         }
 
         public async Task ProductServiceUpdate(Product product)
         {
-            _productDomainServiceValidation.ValidateHorti(product);
+            _updateProductDomainServiceValidation.ValidateHorti(product);
 
             await _productRepository.UpdateProduct(product);
         }
